fixed async example
