+++ package.json
@@ -20,7 +20,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/effektif/react-mentions.git"
+    "url": "https://github.com/signavio/react-mentions.git"
   },
   "keywords": [
     "react",
@@ -30,9 +30,9 @@
   "author": "Jan-Felix Schwarz",
   "license": "BSD-3-Clause",
   "bugs": {
-    "url": "https://github.com/effektif/react-mentions/issues"
+    "url": "https://github.com/signavio/react-mentions/issues"
   },
-  "homepage": "https://github.com/effektif/react-mentions",
+  "homepage": "https://github.com/signavio/react-mentions",
   "devDependencies": {
     "nwb": "0.21.x",
     "prettier": "^1.10.2",
